@@ -9,24 +9,23 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@datadog/browser-logs": "^5.10.0",
+    "@datadog/browser-logs": "^5.23.3",
     "@livekit/components-react": "2.4.3",
     "@livekit/components-styles": "1.0.12",
-    "@livekit/krisp-noise-filter": "^0.2.5",
-    "livekit-client": "2.5.0",
+    "@livekit/krisp-noise-filter": "^0.2.7",
+    "livekit-client": "2.5.1",
     "livekit-server-sdk": "2.6.1",
-    "next": "14.2.5",
-    "next-seo": "^6.0.0",
+    "next": "14.2.7",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "tinykeys": "^2.1.0"
   },
   "devDependencies": {
-    "@types/node": "20.14.13",
-    "@types/react": "18.3.3",
+    "@types/node": "20.16.3",
+    "@types/react": "18.3.5",
     "@types/react-dom": "18.3.0",
-    "eslint": "9.8.0",
-    "eslint-config-next": "14.2.5",
+    "eslint": "9.9.1",
+    "eslint-config-next": "14.2.7",
     "source-map-loader": "^5.0.0",
     "typescript": "5.5.4"
   },
